BUGFIX: Fix error in the switch validation method.
In the Game_Switches_Class::isValidSwitch(int) method, comparison
for the upper bound (< switches.size) was wrong since switch IDs ranges
from 1 to switches.size inclusive.
This caused easyrpg-player to exit with an error message if the game
made use of the last switch.
Replacing the < comparison operator with <= fixes that.