Fix #358.
Changes the PNG reader for paletted images to use the index of a
pixel to decide its transparency, rather than its color. This caused
pixels with non-zero index to be transparent if their index's color
happened to be the same as index 0's color.

Unfortunately, the indexed color -> RGBA conversion was being done
in libpng, so EasyRPG never saw the indices to begin with. So it
requires fairly invasive changes. For the sake of symmetry, I pulled
out functions for non-indexed images too. Not very pretty.