Avoid allocating temporary array.
When writing a row of pixel data, instead of using a temporary
buffer of width w to hold the indices, put them in the last
w bytes of the row of pixel data. Then the index->RGBA loop looks
like

    read nth index (from offset 3w+n)
    convert to color
    write color to nth pixel (offsets 4n..4n+3)

To check we don't overwrite a byte holding an index we'll later
need, we need to know that the offset of the last byte written,
4n+3, is less than the offset of the next byte read, 3w+(n+1).
4n+3 < 3w+(n+1) is equivalent to n < w-2/3, which is true since
n <= w-1.