Update events' move routes after their commands
Checking if a move route has ended is now done after the move route for-
loop. This used to be before the for-loop to make sure the move route
was cleaned a frame after it had ended. That way the ProceedWithMovement
would let the event script keep running at the correct frame.
However, things get complicated if the move route and the
ProceedWithMovement commands are in different events. Let's call these
events A and B respectively. If A has a smaller ID than B, it turns out
that B passes through the ProceedWithMovement in the same frame A's move
route ends. But A would need another frame to surpass a
ProceedWithMovement.
That means that A's route is updated after its commands