autotools: (re-)generate the manual page automatically
If a2x (asciidoc) tool is available, build the manual page for git
checkouts and rebuild when necessary for distributed files as well.
If a2x is not available, print a warning.
If using a distribution archive (i.e. manual page is included) or the
a2x tool is available, then install the manual page.