Follow instruction to build and launch the project
The instructions :

Warning:Native C/C++ source code is found, but it seems that NDK option
is not configured.  Note that if you have an Android.mk, it is not used
for compilation.  The recommended workaround is to remove the default
jni source code directory by adding:
 android {
    sourceSets {
        main {
            jni.srcDirs = []
        }
    }
}
to build.gradle, manually compile the code with ndk-build, and then
place the resulting shared object in src/main/jniLibs.