Tweak event movement routes
Push the move routes methods down in the Game_Event class.

If an event can't do a step, it won't turn to the direction of the
failed step. This is done to make the movement look more natural.

In MoveTypeRandom: After using a dynrpg plugin to print the values of
stepFrameCounter (in liblcf: stop_count) and frequencyTimer (in
liblcf:max_stop_count), I've simulated the randomness of these two
variables. After about 1000 steps, 38% of the steps were forward steps
in RPG_RT, while 70% in Player. This has been reduced to 40% in Player.