Battle interpreter changes: Only search for runnable pages when the scene says that it is a good time to run the interpreter at certain points. Only reset "X used command" pages when that actor had another turn.
In the old behaviour the interpreter always looked in each update-loop if a page can run. This makes sense for the map but not at all for the battle because the only time when pages run is when the turn starts.
The interpreter behaves now correctly in OFF (uses conditional branch) and in Ara Fell (uses event page condition) for "X used command".
The timing when the page runs is now also correct. RPG2k3 behaviour should be perfectly replecated now.

Fix #488
Fix #983