Delay teleport when a parallel interpreter calls Show/EraseScreen.
RPG_RT does not yield the interpreter when doing Transitions compared to our code which does the transition in the graphic update.
Due to this a pending teleport is executed directly after a Show/EraseScreen due to our yield. This means it is executed too early and events following a Show or Erase screen are not executed after a map change (which is wrong).
By inventing an additional flag the teleport is delayed (= not executed) when the interpreter yields due to a Show/EraseScreen, continuing the event execution on the same map as expected one frame later.

It's possible that this causes some timing issues but this flag is only relevant for the mentioned corner case.
Show/Erase on the main interpreter will not toggle it (typical Parallel vs. Main inconsistency in RPG_RT).

Fix #1032