Allow replaying input logs
A new Input::Source is added that reads pressed_buttons from the
log file written by --record-input. Another command-line arg is added
to select the log file for replaying.

If --seed is also used, this *should* give perfectly reproducible
runs of a game (eg. for automated testing, etc.). Main caveat is that
resetting the game (with F12) does not go through the Input
module so it doesn't get reproduced.