Fix BG bug introduced in fc546f70
The assumption that a map had finite height caused rendering issues.
The y-direction doesn't need to be considered, but it was included
to make the map representable as a Rect. When the map loops
vertically, this is impossible and caused the bug.

Solved by inlining the call to Rect::Adjust and removing all handling
of the y-direction.