Add fonts compatible with RMG2000 and RM2000 for unofficial translations
Japanese fonts have a few quirks that don’t need to be reproduced
in other languages. So, it makes sense to have separate Japanese
fonts for official translations, and other fonts for unofficial
translations.

The major differences of Japanese fonts are:

— ` (grave accent) is opening quote,
— ' (straight apostrophe) is a closing quote,
— " (straight double quote) is a closing double quote,
— Cyrillic and Greek characters are double-width,
— ellipsis is in the middle of the line (also useful for Chinese
and maybe Korean).

For RMG2000, a font similar to Shinonome is used (but without the
peculiarities described above).

For RM2000, a new font is added. Like Don Miguel's RM2000, it has different
style for Cyrillic and Latin, Cyrillic is bigger than Latin. For
Cyrillic, I've used a 'straight italic' style (so-called Bulgarian cyrillic).
Its look-and-feel is similar to Don's RM2000 font.