Use ttyp0 for RM2k3E
The original idea behind using Japanese fonts for the official
English release was that it used MS Mincho/MS Gothic. However,
upon closer inspection I've noticed that the official releases
use EDITED MS Mincho and MS Gothic (in the official RM2kE, they
are called RPG2000 and RPG2000G), and those edited fonts use
a European … and not Japanese). So, I think it’s safe to lump
them alongside with the unofficial translations and use ttyp0
for them.