Move error message and <?> glyph from find_glyph to find_fallback_glyph
If the glyph is not found in just one font, but is found
in another (e.g. not found in Mincho, but found in Gothic),
this is not something that needs reporting.

Also, moving the <?> handling into find_fallback_glyph
makes other glyph-finding functions cleaner since they
only need to check for NULL.