Fix #1160; regression from 7f83a3bf
This was caused by a bad refactor. Previously we had

    to scroll up:
      parallax_y += min(dist, dist * A/B)
    to scroll down:
      parallax_y -= min(dist, dist * A/B)

I refactored to handle scrolling up as a special case of scrolling
down with a negative distance, but did not update the minimum, which
changes to a maximum on negation (-min(x, y) = max(-x, -y)).

This is now

    to scroll down:
      parallax_y -= closer_to_zero(dist, dist * A/B)

which is correct in both cases (the two arguments have the same sign
and therefore closer_to_zero(x,y) is min(x,y) if they are positive
and max(x,y) if they are negative.)