GetTimer returns now the remaining seconds (rounded up) instead of the frames.
This fixes a bug when code checks for "Timer >= 1s" followed by "Timer <= 0".
Because EasyRPG checked the frames both tests failed for the frames 1-59 (less then 1s, more then 0s).
The value is now rounded up to next second, to match RPG_RT behaviour.

Thix fixes another bug in CommandControlVariable: For Timer reads the variable was set to the frames, not to seconds.

Fix #1165