Increase accuracy for cycle movement
Events should move in the direction they are facing if possible
rather than retaining the direction to cycle in their state.
The delay between movements was also matched up to RPG_RT.
Particularly in the case that the event is between two obstacles
on either side and can't move, the interval in frames between
the times it turns should be correct now.