Emscripten: Don't finish the teleport until all assets that will be required by the map are downloaded.
The purpose is to make the fade-in look correct. Currently only the ChipSet
is "important". Which means other assets are missing and will "pop-in" after
the fade-in is finished which is wrong and looks ugly.

Disadvantage is that map loading will take slightly longer because the Player
waits for all file transfers. Still faster then MV, though ;)