Hopefully make handling of classes, class changes, and battle command changes better. [P]
This fixes handling of classes, with the new consideration that "changed_class" is actually more "changed_battle_commands", and what this implies.
Specifically that any code relying on changed_class's old meaning is based on a lie.
It also fixes handling of battle command changes, now that '-1 in the array means check LDB' is pretty clearly false

(But '-1 as a LSD class_id means check LDB' is true, which suggests possible miscommunication somewhere.)