SetVehicleLocation actually supports changing the map without doing a transition while the party is boarded but only when the target map has an event with the same event ID.
Because this sounds more like a bug than intended behaviour this was implemented using a normal teleport and without a check if the target map has a valid event. Doesn't match the RPG_RT behaviour due to the transition but this bug is strange and only used by a few VIPRPG games, not worth implementing perfectly.