Vehicles notify now the Player when unboarding is finished.
Before the information was incorrect for one frame which caused an issue with Event command "Conditional Branch -> Is Driven" in combination with "Set Vehicle Position" because the Player was moved together with the vehicle even though it was reported as not being driven.

Fixes landing with the broom in Witch's Heart. Thanks to vgperson for figuring this out.

Fix #1349