"4.5: RM2000 + RM2003: The "DECISION" sound when pressing the Attack command should only sound once, not twice."
Solution: Scene_Battle::AttackSelected already has the "DECISION" sound. Therefore, it's not needed for it to sound when the command is selected. We eliminate it then for the "Attack" command in the Scene_Battle_Rpg2k::CommandSelected and Scene_Battle_Rpg2k3::CommandSelected (for the other commands (skill, items, defend, etc.) it's still kept).