"2.4: RM2000 (not tested in RM2003): If you have the "special skill name" (to give your character a different name to the skill command) checkbox unmarked, it should show the default skill name instead of what is inside."
Solution: Instead of checking whether the skill rename is empty, it checks whether the rename_skill checkbox was clicked or not (after all, if it's clicked, you can put any name, even an empty one).
For this, we need to add in Game_Actor::IsRenameSkill, which access to that attribute and was missing.