Fix Issue ( #1371 ): "4.2: RM2000 + RM2003: There should be white flashes when a battle starts between the battle sound and the TransitionOut. If we look the video in slow motion we can see it's not 1 flash, but two."
Solution: Add an Transition::AppendBefore function summoned right after the Transition::Init in Scene_Map::TransitionOut when there is a battle. AppendBefore has as arguments the color of the flash, the frame duration of each flash, and the number of iterations. The modifications make that when the Transition is drawn, first it generates those flashes in the frozen screen, and then it starts the selected transition.