"2.12: RM2000: Using an object that restores SP displays the message with the amount -1 (it recovers the normal SP though)."
Solution: Fix an error where in the GetHpSpRecoveredMessage it took GetAffectedHp instead of value (which changes for HP or SP). It returned -1 because Ethers only cure SP.

By checking this, it's also been fixed "GetHpSpAbsorbedMessage" where it printed Data::terms.health_points instead of the input parameter "points" (which means, if you absorb SP, it would still say it was HP in the message).