3.3 + 3.5 (together because they have some common code):
"3.3; RM2000: An absorb attack shouldn't make the enemy flash and generate "Damage" sound."
"3.5: RM2000: If a skill decreases stats that are not HP, the enemy shouldn't flash and the "Damage" sound shouldn't play."

Solution for 3.3: Put a condition so it doesn't flash or generate sound. For this:
1. We create a public bool function IsAbsorb() in AlgorithmBase which returns the bool "absorb", so it can be used outside.
2. Change every instance where absorb is used to put IsAbsorb (this is done for consistency since other parameters use the same).
3. To not generate the sound: in AlgorithmBase::GetResultSe, if the attack is absorb, it returns NULL.
4. To not generate the flash: in Scene_Battle_Rpg2k::ProcessBattleAction, in the BattleActionState_Result, when the enemy flashes, it should be put as a condition that the attack shouldn't be absorb.

Solution for 3.5: Put a condition so it doesn't flash or generate sound. We know if GetAffectedHp is -1, HP is not a parameter, so we'll use this check:
1. To not generate the sound: in AlgorithmBase::GetResultSe, if the successful attack doesn't affect HP, it returns NULL.
2. To not generate the sound: in Scene_Battle_Rpg2k::ProcessBattleAction, in the BattleActionState_Result, when the enemy flashes, it should be put as a condition.

Also added that the attack shouldn't be IsPositive(): it shouldn't flash if the enemy is healed.