Improve RandomBlock Draw performance.
Problem: When Drawing the RandomBlockFunction, by the end of the function it was still in a "for" with 4800 Blit operations, which could be costly the more they were printed.

Solution: Create a BitmapRef called random_block_transition = transparent and an Int called current_blocks_print = 0. Each time the Draw function is called for a RandomBlockTransition, it checks the number of blocks to be printed from the previous time Draw was called to this one, and they are printed in random_block_transition (which means, in cases of constant low Draws, it only has to print a few Blits). Then the first screen and the second random_block_transition screen is printed. It helps performance during the Draw operation.