Objective: Implement the #1377 feature from RPG_RT:
"When a skill (with animation attached) is directed towards an ally target or ally group (enemy skill, or ally skill that heals), instead of playing the graphic animation, it plays only the sound. This is fine, but there is more than that: It only plays the sounds of the 20 first frames of the animation. If a sound is on frame 21 or further, it won't play. Also, the waiting time will always be around 15-20 frames, no matter its actual length (more or less)."

Solution:
 - Include a cutoff frame where, from the on, any Sfx operation (sound, flash or shake) isn't executed. We include this variable as an attribute in battle_animation and as an input argument for the constructor (mostly applied to BattleAnimationBattlers). Then we use then for any subsequent function made to invoke Battle Animations that (in our case) could summon SoundAnimation. Therefore, in Scene_Battle_Rpg2k.cpp, when it invokes that function, it puts 20 as the cutoff point.
 - Also, since the wait time for this animation doesn't matter, what we do is add a "ShouldOnlySound" function available for Scene_Battle_Rpg2k.cpp, so in case it's a sound, it doesn't have to wait until it's finishes, it just waits the normal GetDelayForWindow().
 - In that case, the wait after any animation is finished has been changed from GetDelayForLine() to GetDelayForWindow(), it's more exact.