"2.15: RM2000: The conditions caused by a weapon attack are different from the ones selected in RPG_RT (i.e. a weapon with Poison and Blindness, causes Death)."
Solution: This happened because when calculating which state it was, they put state_set[i], which is not the answer because state_set is a set of booleans that mark if the state i is checked or not. Therefore, because the conditional before said it's checked, state_set[i] => i + 1 (because the ID starts with 1 instead of 0).

Also, since an enemy isn't inmediately in Death condition when taking the Death Message, and it's the only message that matters, we change "GetTarget()->GetSignificantState()" for "ReaderUtil::GetElement(Data::states, 1)" (always Death).