2.6 + 3.7
"2.6: RM2000 (issue #1102 ): When an enemy dies, a message should be displayed right after the attack inside the same message."
"3.7: RM2000: If an enemy dies with an attack that takes away many stats HP included (or puts conditions), the enemy should die right after the HP damage is done, and the other messages are ignored. This also includes absorb attacks."

Solution: We push that messages after the damage if the enemy is dead, and it skips the rest of result messages.