EXE ExFont extraction.
While this code probably isn't perfect, it's simple.
If no ExFont file is present as a known image, one will be written to the save directory.
NOTE: This occurs even if one is already in the save directory that is not recognized as a game directory image.
This quirk is so Emscripten-based game updates work properly.
Also, adds support for loading ExFont.bmp from the save directory.