Blit Weather images without showing the transparent pixels:
The reason why they were printed was because Weather uses ToneBlit over itself, which the way bitmap.cpp deals with means it shouldn't skip pixels with alpha 0. To solve it, we add to ToneBlit an optional input bool "check_alpha" (default false), which will check that parameter even if "&src == this". We apply it to the rain and snow (check_alpha = true).