Adding quick TransitionIn to map after Menu/Save/Name/Shop/Load calling
Problem: "The transition between one of those menus to the previous map should be a quick FadeIn, but it fades in with the default Scene_Map.

Solution: Add a "transition_menu" bool in Game_Temp that is activated everytime a menu is called and checked in Scene_Map::TransitionIn, where it's deactivated and it calls Scene::TransitionIn (quick fadein).