Stop sounds (OFF) / (...) only when CommandPlaySound is called.
Problem: An option where, if an (OFF), Empty or filename in parenthesis is called, makes all the sound effects stop (SE_Stop). This also happened in instances outside "CommandPlaySound" (where the creator puts it intentionally), for example in Battle sound animations for effects that include anything but a flash, and generated sound cuts everywhere.

Solution: Add in SePlay and OnSeReady an optional input bool "stop_sounds" (by default "false") that only stop all sounds if it's true, with the OnSeReady input being dependent of the SePlay input. We also put SePlay in CommandPlaySound with stop_sounds = true, the only case where we want it to.