"2.36: RM2000: If a skill causes a state that the target already has, it doesn't try to do the to_hit condition chance, it always shows the "already" message. If a skill cures a state that the target doesn't have, it's not included in the "condition" list (the skill could miss)."
Solution: Before it checks randomly if the attack missed or not, it checks whether the character is already in that state. If it is, then push in conditions inmediately. The same way, if the skill is positive but the target doesn't have the skill, it misses the to_hit chance to skip to the next condition as if it failed.

We also remove the condition that a healing condition doesn't need to do the to_hit rate check. It needs to.