"2.37: RM2000 + RM2003: If a skill resurrects a character but doesn't have the "affect_hp" checkbox, the effect value will restore the character's %HP (100 or more will restore the character completely). If other stat checkbox are selected, they still gain the same amount of absolute effect value."
Solution: If "affect_hp" isn't put and the condition restores Death, then change the HP to cure the percentage of the target's HP. Since we need the conditions saved to know whether it cured Death or not, this has to go after conditions (way after the parameters). So we need to preserve the "effect" value, which is why the local variable "effect" has been put at the beginning of the function.