"2.40: RM2000: Create animation of enemy self-destructing, similar to Death but their image stretching a little bit."
Solution: Adding a new AnimationState_DeadSelfDestruct that stretches the zoom image at the same time it fades out. We prepare it so the values reset when other animation is done.

Also, since it's possible for the enemy to have a ConditionHeal message, we want the initial flash (AnimationState_SelfDestruct, without the "Dead") to appear before, and then when it applies the SelfDestruct, to play the animation:

 - We create in Game_BattleAlgorithm a virtual function (overrided by SelfDestruct) called GetSourceAnimationStateApply.
 - In Scene_Battle_Rpg2k we execute that function right after Apply only if it's not NULL.