"2.41: RM2000: If a battler loses an item or SP in the middle of a turn before their turn is in, and they can't throw that action, their turn becomes NoMove."
Solution: Change UpdateBattlerAction to account for these cases and update to NoMove if it cannot make that action. Since we need to know the SP or item, we add those functions to Game_BattleAlgorithm /GetItem, since it's shared by Skill and Item, it's a virtual function in AlgorithmBase. SpCost however is specific only to Skill).