"2.42: RM2000: Implement a Random Agility modifier when deciding the execution order of the battlers (-10%, 10%)."
Solution: After doing several tests, aproximately the way it works is by varying the Agi for the execution between (-10%, 10%), with 1 value more or less for each extreme (for example, if you have two characters with Agility 10 and 8, the second one still has a few chances of going first).

Since we can't include in the sort function the random value (it would break the "Restrict Weak Ordering" property), we need to set them before the sort function. We add in Game_Battler a function to set randomly the Agi value and a Getter to change it inside the Sort function.