"3.15: RM2000: The Defense resistance should be applied from the beginning of the turn to the end, regardless of when the turn of the battler comes."
Solution: Since we need the defense to be applied at the beginning of the turn, and we don't want to interfere with 2k3, we create a new Game_BattleAlgorithm function called ApplyFirst, which is executed at the beginning of the turn for all battlers after the order has been decided. This function is overrided by Defend.

We also need to cancel the defense value at the end of the turn for all the battlers.