"3.17: RM2000 + RM2003: A Weapon's critical chance is added to the critical chance's natural possibility of the actor."
Solution: The way the critical weapon chance worked before was proportional to the current critical chance of the battler. That means, if the battler has 0, it cannot have critics, when in reality it should. It should increase the possibilites based on the percentage that is still not a critical hit. That's why, instead of "crit_chance * weapon->critical_hit", it should be "(1.0 - crit_chance) * weapon->critical_hit"