If actor = "Prevent Terrain Damage", they don't receive damage.
Problem: Every character receives damage in Terrain that does it regardless of that protection. That protection only affects that single character, not the group.

Solution: We create Game_Actor::PreventsTerrainDamage which check if they have a "Prevent Terrain Damage" protection. Then we check this when terrain damage is going to be done, and if they don't, suffer the damage for that character alone (and generate the red flash).