Make Game_Actor::IsItemUsable account for hero classes in 2k3.
Solution: If it's Rpg2k3, consider the hero class with similar conditions to the actor_set.

An item class_set has as first element [0] "None", which means this happens if the actor has no class (GetClass returns nullptr). This also means the first class is in position [1] instead of [0], like an one-offset. This is only appliable to this particular "class_set".