Operations to UseSkill with source; GetHighestLeveledActor for menu special items.
Problem 1: Skills used in the menu didn't take in factor attributes and the source's ATK and INT depending of the skill. Solution: Add a Source to the function and when a skill is used, pass the source actor. We need to add GetAttributeMultiplier to the Actor from Game_BattleAlgorithm to make this operation too. We also take in account whether a character is resurrected with skill->affect_hp being false (the base value will cure HP percentage in that case), using "cure_hp_percentage".

Problem 2: When using a special item, the parameters it takes are from the highest leveled character in the group. In case there are many, it takes the first one. We pass Game_Party::GetHighestLeveledActor as parameter.