BeginJump calculating the right direction at the end.
Problem: In a Jump, all the Turns commands are ignored. Unless the character's view is fixed somehow, the character will always look to the direction they moved the furthest to (with prevalence with Down/Up directions, unlike diagonal movement). If not, it looks at the same direction it started before the jump, regardless of the Turns it made during the jump.

Solution: Add, if the character doesn't have a fixed direction, look at the direction they've gone the furthest. This code must be executed when the move failed but the command isn't skippable so it looks at least in that direction. Otherwise, look at the same direction it started before the jump.