When movement finishes, frequency wait is 0.
Problem: In RPG_RT, when a movement ends (without loop) the frequency wait isn't executed (so once their movement ends with "Wait to end All Movements", it's right after it's stopped). In EasyRPG, there's still a little wait.

Solution: If the movement ended and it's not a loop, turn the max_stop_count to 0.