Not stopping Loop+Skippable last movement if move_failed.
Problem: In RPG_RT, if the movement is loop but skippable, if it cannot execute one movemen and it's the last, it still starts again and tries until it does it. In EasyRPG, their movement is interrupted.

Solution: The function that stops the movement, put as a condition "if (!(active_route->repeat && active_route->skippable))" just for this specific case. The movement will still stop in other cases (overwriting other movement, the "Stop All Movements" command, etc.).