Create Sprite Loop Duplicate graphics with the right Z position.
Problem: When a sprite is drawn, it's only drawn in their position and is not drawn in the other extreme when in a loop. If we add a draw for those conditions, in the case of vertical loop, the duplicate on the other extreme of the screen will be drawn at the same time, therefore being behind or in front of some graphics that should be the opposite.

Solution: To solve this problem, we need to draw those vertical loops before or after the rest have been drawn in the right order. So we need duplicates that have a different X, Y and Z positions and being inserted and deleted from the drawable_list. Because duplicating the Sprite class would be very heavy in memory and performance (because of the bitmap), we create a new Sprite class:

 - SpriteClone: This function is a subclass of Sprite* that inherits GetX, GetY and GetZ, and its constructor asks for those values plus a reference to a sprite.
 - Its function: When the "Draw()" function is called, it calls the "Draw" of its Sprite reference again. Before, it changes its X and Y values and after the draw it puts them back. This way it redraws the function with the right XY parameters by only being a reference.
 - The Z value will determine when it will be drawn.
 - GetType will be a new type "TypeClone" for consistency.
 - We add them to PlayerLib.vcxproj and PlayerLib.vcxproj.filters

To apply this parameter, we need the sprite to create those drawables and remove them. We want to insert everything in Draw() logic since it only concerns drawing. Therefore, for Drawables we create two functions: CreateClones and RemoveClones:

 - CreateClones creates a list of SpriteClones that reference the Sprite itself. Graphics::Draw will call this function before the ordering and draw starts. Most functions will have this empty, except Sprite_Character, which we'll define later.
 - RemoveClones deletes those SpriteClones and their data (and automatically form the Drawable_list). Graphics::Draw will call this function after the draw ends. The function that calls clones must be able to eliminate them.

With this data, we override those two functions in Sprite_Character and Sprite_AirshipShadow, creating three shared_ptr of SpriteClones:

 - Sprite_Character::CreateClones will create clones under these circunstances:
    1. If the Game_Map is an HorizontalLoop with 20/21 size, and some part of the graphic wouldn't be shown in one extreme, it creates a SpriteClone with the same Y and Z, but X on the other side of the screen.
    2. If the Game_Map is a VerticalLoop with 15/16 size, and some part of the graphic wouldn't be shown in one extreme, it creates a SpriteClone with the same X, but different Y and Z. The GetScreenZ is calculated automatically by the Game_Character function, which also forces us to change its Y value temporally (and return it back) to get the right equation.
    3. If the Game_Map has both conditions, it creates a third SpriteClone that moves both X, Y and Z. Therefore, in a corner there can be 4 instances of the sprite drawn.

 - Sprite_Character::RemoveClones will reset the sprite clones.

This draws the sprite but there is a bug with the Z value in Game_Character: GetScreenZ calculates itself based on GetScreenY. This value and GetScreenX however have some loop conditions that alter the value in a way that drastically alters the values (same with jump in Y, which shouldn't be used for calculating the Z position). We need to ignore these when GetScreenZ is calculated. For this:

 - We add in Game_Character::GetScreenX and GetScreenY an optional input argument called "ignore_loop" (default false). If it's true, then it ignores those operations, so it will always calculate the exact Y position. Then, only for GetScreenZ, when it calls GetScreenY, we put "true" as an argument (all the other functions that use it use false by default).

* It's Sprite instead of Drawable because Drawable doesn't have X and Y functions.
** In PlayerLib.vcxproj.filters, fix an error when adding Transition.cpp.