When entering in boat/ship, not synchronize vehicle until the hero is inside
Problem: In RPG_RT, the character moves to the vehicle, then disappears and then the vehicle synchronizes with the hero's movement. In EasyRPG, the vehicle synchronizes before the character has moved towards it, so it teleports and moves even then.

Solution: This happens because the function SyncWithPlayer, which only Game_Player should execute once it's boarded inside, it's used every turn in Game_Vehicle needlessly, rendering the other call useless. It's taken away.