Vehicle follows the hero's direction when entering.
Problem: Just as described in #1403 when entering a vehicle the sprite is facing left but the hero is facing the direction in which entered. This problem disappears once you move, but not before. Out of the three solutions after talking, we've decided to apply n.2 (vehicle follows hero's direction) until we find a game that uses n.1 to its advantage.

Solution: Change the vehicle direction to the character's direction instead of the other way around.