Associate translation from frame to 0.1s to Game_Interpreter_Map.
Problem: Flashes, Tints and Shakes are always translated from frames to tenths (0.1s) except in battler sprites, which is a problem for battle animations that send the same amount of frames for both. This made screen flashes in battle being longer and battler flashes being shorter.

Solution: To be consistent, since Game_Interpreter is the only that takes "tenth" values, it will be the manager of translating from frames to tenths. All Game_Interpreter operations will do the "DEFAULT_FPS / 10" instead of Game_Character and Game_Screen. And since now it works by frames, the Game_Player red flash when damaged terrain is 6 (equivalent to 0.1s) instead of 1.