Jump: make the sprite frame the middle one (idle) if possible.
Problem: In some cases, you can make a jump while the character is still on a movement position instead of the idle.

Solution: Create a function IsMiddleFrameAvailable() which checks if the character animation is one that allows changing it to the middle frame. If it does, it forces that frame.