Avoid a centered sprite's shake while slow scroll/movement.
Problem: When a slow (1/4, 1/8) scroll or movement is happening with a centered sprite, the sprite starts shaking.

Solution: This happens because the function GetScreenX and GetScreenY, by dividing the result by TILE_SIZE, it floors a few numbers that shouldn't, and that makes some conditions in UpdateScroll go like it's lag. To solve it:
 - We create GetRealScreenX and GetRealScreenY, which are like GetScreen except they multiply every parameter by TILE_SIZE, to get the right Screen parameter without being affected by the flooring of the module in the divisions, while still getting the right screen value.
 - In UpdateScroll, in the checks that determine whether to move the camera with the remaining_step, we change "GetScreenXY > center_xy" for "GetRealScreenXY > center_xy * TILE_SIZE" in all cases. This solves the problem.