Move enemy's Escape animation from GetSourceAnimationState to GetSourceAnimationStateApply.
Problem: AnimationSource is executed at the beginning of the action before ConditionHeal. Which means, if it shows a condition message, the enemy's graphic will disappear before the action and Escape sound appears.

Solution: Parting from the GetSourceAnimationStateApply function implemented in 2.40, change it for the enemy. We add an "Escape" animation for the conditionheal in GetSourceAnimationState.