Give Battle priority over Menu; Disable all calls when one is called.
Problem 1: If you press menu at the same time a battle starts, the menu will take priority and will cancel the battle. We fix this by putting the battle first.

Problem 2: Despite this, if you still press menu at the same time, after the battle ends it calls menu. That is because the bool "menu_calling" is still active. To fix all of this, we create a function DisableCalls that disables all the menu calling bools (except "battle_calling" which is only disabled after the battle ends), and call it in all "CallX" functions.