Inmediately start call command when called by an EventCommand.
Problem: In RPG_RT, when a call menu/shop/etc or battle is done by an eventcommand, it is called inmediately regardless of whether the player is still moving. In EasyRPG, it waits until it's over.

Solution: Add a Game_Temp::inmediate_call that is activated when an EventCommand is called. Then, put as a condition in Scene_Map to check the calls if it's active, and disable it in DisableCalls.