Proceed with movement vehicle fix
If you try to move a vehicle which isn't on the map,
ProceedWithMovement will not wait for this to complete.
Instead the movement stays queued forever until you enter
the same map as the vehicle.

This also works if you use SetVehicleLocation() after the
move command to put the vehicle on the same map as you.

Once the vehicle is on the same map as the player, the movement
will immediately execute.

Fixes an issue where the game would hang in HH3 when
ProceedWithMovement was used.

To see the bug before this commit, setup the following event code

```
Move Vehicle Airship: Up
ProceedWithMovement
Teleport to map with airship
```

In RPG_RT you'll teleport, and then upon arriving the airship
will move. In player the game will hang.