Fix airship descent mid-move
Part of #1429. RPG_RT allows you to land the airship
from an event while the player/airship is moving as part of a move
route.

In RPG_RT, this causes graphical glitches as the descent causes
the airship sprite to warp to the tile and change direction before
finishing its walk animation. This set of changes fixes those
graphical glitches in Player while still allowing the same
event behavior to occur.