Use our own processing even for non-transparent PNGs, otherwise libpng makes transparent pixels black.
Fix #1631

That code was created as part of e72fe478 because libpng gives not enough control when the transparency is not palette index 0. Seems the same lack of control happens with non-transparent images: libpng still makes the transparent pixel black.
This uses now the same codepath as for transparent which has shown in 4 years how rock-solid it is.