Debug menu - don't interrupt other scenes.
Imagine a very long cutscene which calls the menu at some point allowing
you to save and prepare your characters for a boss battle. Now suppose
you hit F9 at exactly that moment. The menu gets canceled in favor of
the debug menu and now you start the battle immediately after.

We should not allow the debug scene if any event called another scene.
We should not allow it to be abused to cancel battles.
It also makes sense for ESC menu to take precedence.