Refactor states handling
States are infliced and healed in the order they happen. In order
to faithfully replicate this, in Execute() we make a copy of our
target and then try to add and remove the states. We keep a
record of what we did, then do it again for real in Apply().
Finally, we replay this record once more to print the messages
in 2k battle.

* Implements the skill reverse_state_effect chunk from #821
* Implements physical skills healing states like normal attacks.

This solution is really bad, but so far there appears to be
nothing better without a major refactor of how we process
battle aglorithms.