Battle2k: Delete battle actions for dead battlers
If a battler is killed, just remove them from the action list,
instead of putting a NoMove action.

This makes the timings and behavior for when an actor is killed before
their turn match when they start the battle round dead.

We can also remove the Exists() checks in the algo processing stages.